Somewhat generalize preservation code for pseudo emission.
Previously, a specific preserve fn needed to be called to preserve live
physregs around a pseudo expansion. Now, the expansion is emitted
naively, but the resulting instruction sequence is examined and compared
against the defs of the pseudo. Any discrepancies are saved and
restored, freeing the pseudo expansion code of having to encode what
exactly needs to be saved and restored.

Presently, only A and NZ are supported, and there's no error checking if
the expansions muck with other regs.