Emit simple prolog, epilog, and lower frame indices.
Presently, only supports 4 bytes of hard stack per function invocation.
The prolog and epilog issue up to 4 PHA/PLA instructions. The intent is
to have the SThs and LDhs instructions expand to TSX; STidx $100+offs,X.

This will work, but it'll be far more efficient to have ST lowered to
PHA where possible, and LD lowered to PLA. To do so, we'd need to detect
when its save to weave the prologue in and out of the basic blocks of
the function.

Eventually, a soft stack will be added to support arbitrary sized stack
frames. The intent is to keep the first 4 bytes on the hard stack; these
then become a priveleged "fast stack" for the most important values.
LLVM already optimizes stack indices so that lower indices are used by
more important values, so little extra is required to organize codegen this
way.