Break up frameaddr into halves.
The old FrameAddr instruction produces only one 16-bit output, which
forces it to be allocated to a ZP_PTR register. However, there's no
direct requirement that the high and low bytes be located in ZP_PTR
until the results are used to load/store. Accordingly, splitting the
FrameAddr instruction into AddFiLo and AdcFiHi allows the results
to be kept in some combination of GPR or ZP locations while final
addresses are being computed.