Use register classes instead of physregs.
Machine scheduling considers physical register uses to be very strong
dependencies, and it's not very well adapted to an environment where
nearly every instruction uses a fixed physreg. Instead, create
one-register classes, and the scheduler and register allocator will
work together to ensure that register pressure is within limits, that
is, that no more than one value is assigned to each physical register.