Add dedicated nocallback node to call graph.
The call graph passes really expect all call sites to be included in the
graph, so to avoid upsetting the apple cart too much, I've added another
null node to the CFG for nocallback calls. This keeps them separate from
regular exernal calls, that way when the calls to external are linked up
with the calls from external, the nocallback node is left out.