Add flag to allow disable 16-byte alignment of large globals.
A bit of code archeology suggest that this was added very early on by
Chris Lattner to speed up a Fortran demo involving arrays of doubles.
The reasoning is missing, but it's almost certainly due to either
caching or SIMD concerns, so it can safely be disabled for the 6502.
This saves RAM and simplifies the assembler, since there are no real
alignment constraints for the system.