Lower C structs to non-byval pointers.
By default, clang lowers C structs to byval pointers, but the
interpretation of byval is quite target-specific, and it doesn't
seem to work very well with GlobalISel. Following the example of
the furthest-along GlobalISel target, AArch64, structs are instead
lowered to non-byval indirect pointers. This produces the expected
behavior: argument copies are only made in the caller, and a pointer is
passed to the copy. This works quite nicely with static stacks, as the
pointer can be to the caller's static stack; such pointers can be
generated with only compiler-time arithmetic.